+++ internal/cliutil/util.go
@@ -45,7 +45,7 @@ func TransactionReceiptToString(receipt *protocol.TransactionReceipt, operations
 	}
 
 	// Build the mana result
-	manaDec, err := util.SatoshiToDecimal(int64(receipt.RcUsed), KoinPrecision)
+	manaDec, err := util.SatoshiToDecimal(receipt.RcUsed, KoinPrecision)
 	if err != nil {
 		s += "\n" + err.Error()
 		return s
